# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing the 3X-UI panel including inbounds, server, and settings.",
        "title": "3X-UI Panel API",
        "contact": {
            "name": "3X-UI"
        },
        "license": {
            "name": "GPL-3.0"
        },
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
...
        "/panel/api/server/getConfigJson": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Get Xray config JSON",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/api/server/getDb": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Download database backup",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/panel/api/server/restartXrayService": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Restart Xray service",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/api/server/status": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Get server status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/api/server/stopXrayService": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server"
                ],
                "summary": "Stop Xray service",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/setting/all": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Setting"
                ],
                "summary": "Get all settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
...
        "/panel/setting/update": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Setting"
                ],
                "summary": "Update all settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "entity.Msg": {
            "type": "object",
            "properties": {
                "msg": {
                    "description": "Response message text",
                    "type": "string"
                },
                "obj": {
                    "description": "Optional data object"
                },
                "success": {
                    "description": "Indicates if the operation was successful",
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}